+++ src/lib/pages/publish-module/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./Footer";
+export * from "./PolicyCard";
+export * from "./UploadModuleCard";
+export * from "./accordions";